+++ starter-code/app.js
@@ -50,4 +50,7 @@ app.locals.title = 'Express - Generated with IronGenerator';
 const index = require('./routes/index');
 app.use('/', index);
 
+const celebrities = require('./routes/celebrities');
+app.use('/celebrities', celebrities);
+
 module.exports = app;
