+++ starter-code/views/celebrities/index.hbs
@@ -6,10 +6,10 @@
       <a href="/celebrities/{{this.id}}">
         <h3>{{this.name}}</h3>
       </a>
-      <button>Delete Celibrity</button>
+      <button>Delete Celebrity</button>
     </li>
   </form>
   <a href="/celebrities/{{this.id}}/edit">Edit</a>
   {{/each}}
 </ul>
-<a href="/celebrities/new">Add a new celibrity</a>
+<a href="/celebrities/new">Add a new celebrity</a>
