+++ starter-code/views/celebrities/new.hbs
@@ -10,9 +10,9 @@
   {{#if errorMessage}}
   <div class="error-message">{{ errorMessage }}</div>
   {{/if}}
-  <button type="submit">Add celibrity</button>
+  <button type="submit">Add celebrity</button>
   {{#if successMessage}}
   <div class="success-message">{{ successMessage }}</div>
-  <a href="/celebrities/index">Back to Celebrities List</a>
+  <a href="/celebrities">Back to Celebrities List</a>
   {{/if}}
 </form>
