+++ starter-code/views/index.hbs
@@ -1,2 +1,2 @@
 <h1>{{title}}</h1>
-<p>Welcome to {{title}}, GO TO: <a href="/celebrities/index">Celebrities page</a></p>
+<p>Welcome to {{title}}, GO TO: <a href="/celebrities">Celebrities page</a></p>
